Move storage impls on Fabric from network/node/ to storage/
