+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/internal/storage/StorageInsertableStorage.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage2.platform.fabric.internal.network.node;
+package com.refinedmods.refinedstorage2.platform.fabric.internal.storage;
 
 import com.refinedmods.refinedstorage2.api.core.Action;
 import com.refinedmods.refinedstorage2.api.storage.Actor;
